more compatibility with ds addons, wheel limit increased, float revpush, instructions beautified

+++ src/nfm/lit/Stat.java
@@ -44,7 +44,7 @@ public class Stat {
 
     public int revlift = 0;
 
-    public int revpush = 0;
+    public float revpush = 0F;
 
     public float simag = 0F;
